Revert "Auth api integration (#22)"
This reverts commit 78ff79712cb959d588acc0cebe88ff03e708714d.
