+++ .gitignore
@@ -1,5 +1,3 @@
 node_modules/
 nodemon.json
 coverage/
-globalConfig.json
-.env
