+++ package.json
@@ -5,27 +5,22 @@
   "license": "MIT",
   "dependencies": {
     "apollo-server": "^2.13.1",
-    "bcrypt": "^5.0.0",
     "graphql": "^15.0.0",
-    "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.19",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.6"
+    "typescript": "^3.9.6",
+    "ts-node": "^8.10.2"
   },
   "scripts": {
     "build": "tsc",
     "start": "ts-node src/index.ts",
     "dev": "nodemon --exec ts-node src/index.ts",
     "test": "jest --coverage",
-    "publish:coverage": "./node_modules/coveralls/bin/coveralls.js < ./coverage/lcov.info",
-    "lint": "prettier --write 'src/**/*.{ts,tsx,js,jsx}' 'tests/**/*.{ts,tsx,js,jsx}'"
+    "publish:coverage":"./node_modules/coveralls/bin/coveralls.js < ./coverage/lcov.info"
   },
   "devDependencies": {
     "@shelf/jest-mongodb": "^1.1.5",
-    "@types/bcrypt": "^3.0.0",
     "@types/graphql": "^14.5.0",
     "@types/jest": "^25.2.2",
-    "@types/jsonwebtoken": "^8.5.0",
     "@types/mongoose": "^5.7.30",
     "@types/node": "^14.0.1",
     "apollo-server-testing": "^2.15.1",
@@ -34,7 +29,6 @@
     "jest": "^26.1.0",
     "mongodb-memory-server": "^6.6.1",
     "nodemon": "^2.0.4",
-    "prettier": "^2.0.5",
     "ts-jest": "^26.0.0"
   }
 }
