--- src/graphql/query/sim.ts
@@ -1,32 +1,0 @@
-import Sim from "../../models/sim";
-
-export const sims = async (_root, args) => {
-  const keyword = args.keyword;
-  const query = {
-    $or: [
-      {
-        title: {
-          $regex: keyword,
-          $options: "i",
-        },
-      },
-      {
-        description: {
-          $regex: keyword,
-          $options: "i",
-        },
-      },
-      {
-        tags: {
-          $regex: keyword,
-          $options: "i",
-        },
-      },
-    ],
-  };
-  try {
-    return await Sim.find(query);
-  } catch (err) {
-    console.log(err);
-  }
-};
