--- src/graphql/type-defs/common.ts
@@ -1,7 +1,0 @@
-const commonType = `
-    type DeleteReturnType {
-        success: Boolean!
-    }
-`;
-
-export default commonType;
