--- src/graphql/type-defs/index.ts
@@ -1,21 +1,0 @@
-import QueryType from "./query";
-import MutationType from "./mutation";
-
-import WorkbookType from "./workbook";
-import SimType from "./sim";
-import UserType from "./user";
-
-import CommonType from "./common";
-
-const TypeDefs = `
-    ${QueryType}
-    ${MutationType}
-
-    ${WorkbookType}
-    ${SimType}
-    ${UserType}
-
-    ${CommonType}
-`;
-
-export default TypeDefs;
