+++ tests/mutations.ts
@@ -2,7 +2,9 @@ import gql from "graphql-tag";
 
 export const ADD_WORKBOOK = gql`
   mutation {
-    createWorkbook(workbook: { title: "History", slides: "slide" }) {
+    createWorkbook(
+      workbook: { title: "History", owner: "13jnafso34", slides: "slide" }
+    ) {
       _id
       title
       owner
@@ -14,7 +16,9 @@ export const ADD_WORKBOOK = gql`
 
 export const ADD_FOLDER = gql`
   mutation {
-    createWorkbookFolder(workbookFolder: { title: "Social Science" }) {
+    createWorkbookFolder(
+      workbookFolder: { title: "Social Science", owner: "13jnafso34" }
+    ) {
       _id
       title
       owner
@@ -24,6 +28,7 @@ export const ADD_FOLDER = gql`
 `;
 
 export const UPDATE_WORKBOOK = ({ id, field, value }) => {
+
   return gql`
   mutation {
     updateWorkbook(
@@ -52,8 +57,8 @@ export const UPDATE_WORKBOOK_FOLDER = ({ id, field, value }) => {
         parentId
       }
     }
-  `;
-};
+  `
+}
 
 export const DELETE_WORKBOOK = (id) => {
   return gql`
@@ -64,8 +69,8 @@ export const DELETE_WORKBOOK = (id) => {
         success
       }
     }
-  `;
-};
+  `
+}
 
 export const DELETE_WORKBOOK_FOLDER = (id) => {
   return gql`
@@ -76,63 +81,5 @@ export const DELETE_WORKBOOK_FOLDER = (id) => {
         success
       }
     }
-  `;
-};
-
-export const ADD_USER = (username, email, password) => {
-  return gql`
-    mutation {
-      createUser(userInput: { email:"${email}", password:"${password}", username:"${username}" }) {
-        userId
-        username
-        email
-        type
-        token
-        tokenExpiration
-      }
-    }
-  `;
-};
-
-export const ADDSIM = (id, title, description, tags, imgUrl) => {
-  return gql`
-    mutation {
-      createSim(sim: {
-        _id: "${id}",
-        title: "${title}",
-        description: "${description}",
-        tags: "${tags}",
-        imageURL: "${imgUrl}"
-      }) {
-        _id
-        owner
-      }
-  }
-  `;
-};
-
-export const UPDATE_SIM = (simId, title, description, tags, imgUrl) => {
-  return gql`
-    mutation {
-      updateSim(simId:"${simId}",updatedSim: {
-        title:"${title}",
-        description:"${description}",
-        tags:"${tags}",
-        imageURL:"${imgUrl}"
-      }) {
-        _id
-        title
-      }
-    }
-  `;
-};
-
-export const DELETE_SIM = (simId) => {
-  return gql`
-    mutation {
-      deleteSim(simId: "${simId}") {
-        success
-      }
-    }
-  `;
-};
+  `
+}
