+++ tests/queries.ts
@@ -1,17 +1,17 @@
-import gql from "graphql-tag";
+import gql from 'graphql-tag'
 
 export const GET_WORKBOOKS = gql`
-  query {
-    workbookViewer {
-      _id
-      title
-      parentId
-      type
-    }
+query {
+  workbookViewer (owner: "13jnafso34") {
+    _id
+    title
+    parentId
+    type
   }
-`;
+}  
+`
 
-export const GET_WORKBOOK = (id: string) => gql`
+export const GET_WORKBOOK = (id:string) => gql`
 query {
   workbook(workbookId: "${id}") {
     _id
@@ -21,14 +21,4 @@ query {
     parentId
   }
 } 
-`;
-
-export const LOGIN = (email: string, password: string) => gql`
-query {
-  login(email:"${email}", password:"${password}") {
-    token
-    userId
-    tokenExpiration
-  }
-}
-`;
+`
