--- tests/utils/createMockServer.ts
@@ -1,3 +1,0 @@
-import { createApolloServer } from "../../src/index";
-
-export default createApolloServer;
